# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "CheckModelAccessRequest": {
        "properties": {
          "cp_ids": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Cp Ids",
            "type": "array",
            "uniqueItems": true
          }
        },
        "title": "CheckModelAccessRequest",
        "type": "object"
      },
      "CheckModelAccessResponse": {
        "properties": {
          "access": {
            "additionalProperties": {
              "enum": [
                "authorized",
                "not_authorized"
              ],
              "type": "string"
            },
            "title": "Access",
            "type": "object"
          }
        },
        "required": [
          "access"
        ],
        "title": "CheckModelAccessResponse",
        "type": "object"
      },
...
      "DownloadProgressCompleteResponse": {
        "properties": {
          "status": {
            "const": "complete",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DownloadProgressCompleteResponse",
        "type": "object"
      },
      "DownloadProgressErrorResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "status": {
            "const": "error",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "error"
        ],
        "title": "DownloadProgressErrorResponse",
        "type": "object"
      },
      "DownloadProgressRunningResponse": {
        "properties": {
          "all_files": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "All Files",
            "type": "array",
            "uniqueItems": true
          },
          "completed_files": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Completed Files",
            "type": "array",
            "uniqueItems": true
          },
          "current_downloading_file": {
            "anyOf": [
              {
                "enum": [
                  "ltx-2.3-22b-dev",
                  "ltx-2.3-spatial-upscaler-x2-1.0",
                  "gemma-3-12b-it-qat-q4_0-unquantized",
                  "qwen3-vl-4b-instruct"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Downloading File"
          },
          "current_file_progress": {
            "title": "Current File Progress",
            "type": "number"
          },
          "error": {
            "title": "Error",
            "type": "null"
          },
          "expected_total_bytes": {
            "title": "Expected Total Bytes",
            "type": "integer"
          },
          "speed_bytes_per_sec": {
            "title": "Speed Bytes Per Sec",
            "type": "number"
          },
          "status": {
            "const": "downloading",
            "title": "Status",
            "type": "string"
          },
          "total_downloaded_bytes": {
            "title": "Total Downloaded Bytes",
            "type": "integer"
          },
          "total_progress": {
            "title": "Total Progress",
            "type": "number"
          }
        },
        "required": [
          "status",
          "current_downloading_file",
          "current_file_progress",
          "total_progress",
          "total_downloaded_bytes",
          "expected_total_bytes",
          "completed_files",
          "all_files",
          "speed_bytes_per_sec"
        ],
        "title": "DownloadProgressRunningResponse",
        "type": "object"
      },
...
      "GpuInfoResponse": {
        "properties": {
          "cuda_available": {
            "title": "Cuda Available",
            "type": "boolean"
          },
          "gpu_available": {
            "default": false,
            "title": "Gpu Available",
            "type": "boolean"
          },
          "gpu_info": {
            "$ref": "#/components/schemas/GpuTelemetry"
          },
          "gpu_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Name"
          },
          "mps_available": {
            "default": false,
            "title": "Mps Available",
            "type": "boolean"
          },
          "vram_gb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram Gb"
          }
        },
        "required": [
          "cuda_available",
          "gpu_name",
          "vram_gb",
          "gpu_info"
        ],
        "title": "GpuInfoResponse",
        "type": "object"
      },
      "GpuListResponse": {
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/GpuDeviceItem"
            },
            "title": "Devices",
            "type": "array"
          }
        },
        "required": [
          "devices"
        ],
        "title": "GpuListResponse",
        "type": "object"
      },
      "GpuMemoryResponse": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "total_mb": {
            "title": "Total Mb",
            "type": "integer"
          },
          "used_mb": {
            "title": "Used Mb",
            "type": "integer"
          }
        },
        "required": [
          "available",
          "total_mb",
          "used_mb"
        ],
        "title": "GpuMemoryResponse",
        "type": "object"
      },
      "GpuTelemetry": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "vram": {
            "title": "Vram",
            "type": "integer"
          },
          "vramUsed": {
            "title": "Vramused",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "vram",
          "vramUsed"
        ],
        "title": "GpuTelemetry",
        "type": "object"
      },
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "LtxDownloadRecommendationResponse": {
        "properties": {
          "cps_to_download": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Cps To Download",
            "type": "array"
          },
          "status": {
            "const": "download",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "cps_to_download"
        ],
        "title": "LtxDownloadRecommendationResponse",
        "type": "object"
      },
      "LtxOkRecommendationResponse": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "LtxOkRecommendationResponse",
        "type": "object"
      },
      "LtxUpgradeRecommendationResponse": {
        "properties": {
          "cps_to_delete": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Cps To Delete",
            "type": "array"
          },
          "cps_to_download": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Cps To Download",
            "type": "array"
          },
          "ltx_model_id": {
            "const": "ltx-2.3-22b-dev",
            "title": "Ltx Model Id",
            "type": "string"
          },
          "status": {
            "const": "upgrade",
            "title": "Status",
            "type": "string"
          },
          "upgrade_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upgrade Message"
          }
        },
        "required": [
          "status",
          "ltx_model_id",
          "cps_to_download",
          "cps_to_delete"
        ],
        "title": "LtxUpgradeRecommendationResponse",
        "type": "object"
      },
...
      "ModelDeleteRequest": {
        "properties": {
          "cp_ids": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Cp Ids",
            "type": "array",
            "uniqueItems": true
          }
        },
        "title": "ModelDeleteRequest",
        "type": "object"
      },
...
      "ModelDownloadRequest": {
        "properties": {
          "cp_ids": {
            "items": {
              "enum": [
                "ltx-2.3-22b-dev",
                "ltx-2.3-spatial-upscaler-x2-1.0",
                "gemma-3-12b-it-qat-q4_0-unquantized",
                "qwen3-vl-4b-instruct"
              ],
              "type": "string"
            },
            "title": "Cp Ids",
            "type": "array",
            "uniqueItems": true
          },
          "type": {
            "default": "download",
            "enum": [
              "download",
              "upgrade"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "title": "ModelDownloadRequest",
        "type": "object"
      },
      "ModelDownloadStartResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "status": {
            "const": "started",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "sessionId"
        ],
        "title": "ModelDownloadStartResponse",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
      "TextEncoderRecommendationResponse": {
        "properties": {
          "cp_to_download": {
            "anyOf": [
              {
                "enum": [
                  "ltx-2.3-22b-dev",
                  "ltx-2.3-spatial-upscaler-x2-1.0",
                  "gemma-3-12b-it-qat-q4_0-unquantized",
                  "qwen3-vl-4b-instruct"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cp To Download"
          },
          "expected_size_bytes": {
            "title": "Expected Size Bytes",
            "type": "integer"
          },
          "expected_size_gb": {
            "title": "Expected Size Gb",
            "type": "number"
          }
        },
        "required": [
          "cp_to_download",
          "expected_size_bytes",
          "expected_size_gb"
        ],
        "title": "TextEncoderRecommendationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenLTX Trainer Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/gpu-info": {
      "get": {
        "operationId": "route_gpu_info_api_gpu_info_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Gpu Info",
        "tags": [
          "health"
        ]
      }
    },
    "/api/gpu-list": {
      "get": {
        "operationId": "route_gpu_list_api_gpu_list_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Gpu List",
        "tags": [
          "health"
        ]
      }
    },
    "/api/gpu-memory": {
      "get": {
        "description": "Live used / total VRAM (MB) for one GPU index.\n\nPolled by the Monitor view so the operator can watch VRAM climb,\ne.g. during sample generation where the run is prone to OOM.",
        "operationId": "route_gpu_memory_api_gpu_memory_get",
        "parameters": [
          {
            "in": "query",
            "name": "index",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Gpu Memory",
        "tags": [
          "health"
        ]
      }
    },
    "/api/models/check-access": {
      "post": {
        "operationId": "route_check_model_access_api_models_check_access_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckModelAccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckModelAccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Check Model Access",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/delete": {
      "delete": {
        "operationId": "route_model_delete_api_models_delete_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Model Delete",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/download": {
      "post": {
        "operationId": "route_model_download_api_models_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Model Download",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/download/progress": {
      "get": {
        "operationId": "route_download_progress_api_models_download_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DownloadProgressRunningResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DownloadProgressCompleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DownloadProgressErrorResponse"
                    }
                  ],
                  "title": "Response Route Download Progress Api Models Download Progress Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Download Progress",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/ltx-recommendation": {
      "get": {
        "operationId": "route_ltx_recommendation_api_models_ltx_recommendation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LtxDownloadRecommendationResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LtxUpgradeRecommendationResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LtxOkRecommendationResponse"
                    }
                  ],
                  "title": "Response Route Ltx Recommendation Api Models Ltx Recommendation Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Ltx Recommendation",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/text-encoder-recommendation": {
      "get": {
        "operationId": "route_text_encoder_recommendation_api_models_text_encoder_recommendation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextEncoderRecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Text Encoder Recommendation",
        "tags": [
          "models"
        ]
      }
    },
...
    "/api/settings": {
      "get": {
        "operationId": "route_get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Get Settings",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "route_post_settings_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Post Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/system/shutdown": {
      "post": {
        "operationId": "route_shutdown_api_system_shutdown_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Route Shutdown Api System Shutdown Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Shutdown",
        "tags": [
          "health"
        ]
      }
    },
...
  }
}